Rename and rescope image metadata checker
Previously we did the same requirements checks for image metadata
before submitting the form (pre-update) and before preview. Doing
this sets a labour-intensive precendent for file attachments (for
which we want to enforce official document type selection), and is
also confusing, since there's no clear way for there to be an issue
with the length of image metadata before preview.

This rescopes the requirements checks for image metadata to be
more specific to the context of the check (pre-update or pre-preview)
and in doing so renames the checker to be more specific.
