+++ config/locales/en/requirements.yml
@@ -44,15 +44,12 @@ en:
         summary_message: "Enter alt text for ‘%{filename}’"
       too_long:
         form_message: "Enter alt text that is fewer than %{max_length} characters long"
-        summary_message: "Enter shorter alt text for ‘%{filename}’"
     caption:
       too_long:
         form_message: "Enter an image caption that is fewer than %{max_length} characters long"
-        summary_message: "Enter a shorter caption for ‘%{filename}’"
     credit:
       too_long:
         form_message: "Enter an image credit that is fewer than %{max_length} characters long"
-        summary_message: "Enter a shorter credit for ‘%{filename}’"
     image_upload:
       no_file:
         form_message: Select a file to upload
